Changing all components to be in modlet form #118 and remove unnecessary test.js files #120
