+++ public/components/player/edit/edit.js
@@ -1,15 +1,46 @@
 var Component = require("can/component/component");
-
 require("bootstrap/dist/css/bootstrap.css!");
 require("can/map/define/");
 require("can/construct/super/");
+var Player = require("bitballs/models/player");
 
-var viewModel = require("bitballs/components/player/edit/viewmodel");
+exports.ViewModel = can.Map.extend({
+	define: {
+		player: {
+			Value: Player,
+			Type: Player
+		}
+	},
+	savePlayer: function(){
+		var self = this;
+		var player = this.attr("player"),
+			promise;
 
-module.exports = Component.extend({
+		if(player.isNew()) {
+			promise = player.save().then(function(){
+				self.attr("player", new Player());
+			}).then(function(){
+				can.dispatch.call(self, "saved");
+			});
+		} else {
+			promise = player.save().then(function(){
+				can.dispatch.call(self, "saved");
+			});
+		}
+
+		this.attr('savePromise', promise);
+
+		return promise;
+	},
+	cancelEvent: function() {
+		can.dispatch.call(this, "canceled");
+	}
+});
+
+exports.Component = Component.extend({
 	tag: "player-edit",
 	template: require("./edit.stache!"),
-	viewModel: viewModel.extend({
+	viewModel: exports.ViewModel.extend({
 		savePlayer: function(ev) {
 			ev.preventDefault();
 			this._super.apply(this,arguments);
@@ -24,5 +55,3 @@ module.exports = Component.extend({
 		}
 	}
 }); 
-
-// promise
