+++ public/components/player/edit/edit_test.js
@@ -1,15 +1,16 @@
 import can from 'can';
 import stache from 'can/view/stache/';
 import QUnit from 'steal-qunit';
-import ViewModel from 'bitballs/components/player/edit/viewmodel';
+import playerEdit from 'bitballs/components/player/edit/edit';
 import Player from 'bitballs/models/player';
 import F from 'funcunit';
 import route from "can/route/";
 import Session from "bitballs/models/session";
 import './edit';
 
-
 import 'bitballs/models/fixtures/players';
+
+var ViewModel = playerEdit.ViewModel;
 
 F.attach(QUnit);
 
@@ -23,9 +24,10 @@ QUnit.module('player/edit', function(hooks){
 	QUnit.module('ViewModel', function(){
 
 		QUnit.test('Tests are running', function(assert){
-		  assert.ok( true, "Passed!" );
+			assert.ok( true, "Passed!" );
 		});
 
+		var vm = new ViewModel();
 		QUnit.test('Can create new ViewModel', function(assert){
 			var vm = new ViewModel();
 				
@@ -37,7 +39,7 @@ QUnit.module('player/edit', function(hooks){
 				vm.unbind("saved");
 				done();
 			});
-			vm.savePlayer()
+			vm.savePlayer();
 		});
 
 		QUnit.test("Create player", function(assert){
@@ -54,13 +56,13 @@ QUnit.module('player/edit', function(hooks){
 					player:playerModel
 				});
 
-				vm.bind("saved", function(){
-					player.id = 1;
-					assert.deepEqual(player, playerModel.attr(),  "New player saved");
-					vm.unbind("saved");
-					done();
-				})
-				vm.savePlayer();
+			vm.bind("saved", function(){
+				player.id = 1;
+				assert.deepEqual(player, playerModel.attr(),  "New player saved");
+				vm.unbind("saved");
+				done();
+			});
+			vm.savePlayer();
 			
 		});
 
@@ -77,12 +79,12 @@ QUnit.module('player/edit', function(hooks){
 					player: playerModel
 				});
 
-				vm.savePlayer()
-				vm.attr("savePromise").fail(function(resp, type){
-					assert.equal(type, 'error', 'fail creation without password');
-					assert.equal(vm.attr('savePromise').state(), 'rejected');
-					done();
-				});
+			vm.savePlayer()
+			vm.attr("savePromise").fail(function(resp, type){
+				assert.equal(type, 'error', 'fail creation without password');
+				assert.equal(vm.attr('savePromise').state(), 'rejected');
+				done();
+			});
 		});		
 
 		QUnit.test("Update player", function(assert){
@@ -100,17 +102,17 @@ QUnit.module('player/edit', function(hooks){
 					player:playerModel
 				});
 
-				//update player info
-				vm.attr("player.name", "Test Player (modified)");
+			//update player info
+			vm.attr("player.name", "Test Player (modified)");
 
-				vm.bind("saved", function(){
-					player.name = "Test Player (modified)";
+			vm.bind("saved", function(){
+				player.name = "Test Player (modified)";
 
-					assert.deepEqual(playerModel.attr(), player, "Player updated");
-					vm.unbind("saved");
-					done();
-				});
-				vm.savePlayer()
+				assert.deepEqual(playerModel.attr(), player, "Player updated");
+				vm.unbind("saved");
+				done();
+			});
+			vm.savePlayer()
 			
 		});
