+++ public/components/player/list/test.html
@@ -1,2 +1,2 @@
-<script src="../../../node_modules/steal/steal.js" main="bitballs/components/player/list/player_list_test"></script>
+<script src="../../../node_modules/steal/steal.js" main="bitballs/components/player/list/list_test"></script>
 <div id="qunit-fixture"></div>
