--- public/components/player/page.stache
@@ -1,7 +1,0 @@
-<can-import from="bitballs/components/edit/" />
-{{#if showList}}
-	<player-list/>
-{{/if}}
-{{#if showEdit}}
-	<player-edit player-id='{route.playerId}'/>
-{{/if}}
