--- public/components/player/test.js
@@ -1,4 +1,0 @@
-// TODO: make the player vs players fixtures play nice
-// import './edit/edit_test';
-import './list/player_list_test';
-
