+++ public/components/test.js
@@ -1,4 +1,6 @@
-import './game/test';
-import './player/test';
-import './user/test';
+import './game/details/details_test';
+//import './navigation/navigation_test';
+//import './player/edit/edit_test';
+import './player/list/list_test';
 import './tournament/test';
+import './user/user_test';
