+++ public/components/tournament/details/details_test.js
@@ -1,6 +1,8 @@
 import QUnit from 'steal-qunit';
-import ViewModel from './viewModel';
+import details from './details';
 import defineFixtures, { tournaments } from 'bitballs/models/fixtures/tournament';
+
+var ViewModel = details.ViewModel;
 
 QUnit.module('Tournament Details ViewModel', {
     beforeEach: function () {
