+++ public/components/tournament/details/test.html
@@ -1,3 +1,3 @@
 <title>Tournament Details Tests</title>
-<script src="../../../node_modules/steal/steal.js" main="bitballs/components/tournament/details/test"></script>
+<script src="../../../node_modules/steal/steal.js" main="bitballs/components/tournament/details/details_test"></script>
 <div id="qunit-fixture"></div
