+++ public/components/tournament/list/list.stache
