+++ public/components/user/test.html
@@ -4,6 +4,6 @@
   <body>
     <div id="qunit"></div>
     <div id="qunit-fixture"></div>
-    <script src="../../node_modules/steal/steal.js" data-main="bitballs/components/user/test"></script>
+    <script src="../../node_modules/steal/steal.js" data-main="bitballs/components/user/user_test"></script>
   </body>
 </html>
