+++ public/components/user/user.html
@@ -1,5 +1,5 @@
 <script type="text/stache" id="demo" can-autorender>
-  <can-import from="bitballs/components/user/create" />
+  <can-import from="bitballs/components/user/user" />
 
   <user-create></user-create>
 </script>
