+++ public/components/user/user_test.js
@@ -1,7 +1,9 @@
 import $ from 'jquery';
 import QUnit from 'steal-qunit';
-import ViewModel from 'bitballs/components/user/view-model';
+import user from 'bitballs/components/user/user';
 import 'bitballs/models/fixtures/user';
+
+var ViewModel = user.ViewModel;
 
 QUnit.module('User: create', {
 	beforeEach: function() {
